* lib/qwik/act-pagelist.rb: Fix bug.
git-svn-id: svn://rubyforge.org/var/svn/qwik/qwik@620 4820488e-9c09-0410-87d0-9d66fe36c344

+++ lib/qwik/act-pagelist.rb
@@ -26,7 +26,7 @@ module Qwik
         dat = Marshal.dump(ar)
         pagelist_path.write(dat)
 
-        running_path.unlink
+        running_path.unlink if running_path.exist?
       }
     end
